Switch to using extracted mbtapi version of api client
https://github.com/kadamwhite/mbtapi is the extracted, library-ified
version of this logic, so it is advantageous to use the standalone
module for consistency and mockability
