+++ knexfile.js
@@ -1,5 +1,5 @@
 // Update with your config settings.
 
-var config = require( './server/lib/config' );
+var config = require( './server/services/config' );
 
 module.exports = config.db.environments;
