+++ package.json
@@ -76,6 +76,7 @@
     "knex": "^0.7.6",
     "lodash": "^2.4.2",
     "lru-cache": "^2.6.5",
+    "mbtapi": "^0.2.0",
     "morgan": "^1.6.1",
     "nodemon": "^1.4.1",
     "pg": "^4.4.1",
