+++ tests/server/lib/config.js
@@ -10,7 +10,7 @@ var proxyquire = require( 'proxyquire' );
 describe( 'makeQueryHandler', function() {
 
   it ( 'loads and parses the configuration file', function() {
-    var config = proxyquire( '../../../server/lib/config', {
+    var config = proxyquire( '../../../server/services/config', {
       'js-yaml': {
         safeLoad: function() {
           return require( '../../mocks/mock-config' );
@@ -34,7 +34,7 @@ describe( 'makeQueryHandler', function() {
 
   it ( 'throws an error if the API key is not present in the config', function() {
     expect(function requireModule() {
-      proxyquire( '../../../server/lib/config', {
+      proxyquire( '../../../server/services/config', {
         'js-yaml': {
           safeLoad: function() { return { api: {} }; }
         }
